mds/MDLog: review and cleanup the accounting of LogSegments
The main goal of this change is to keep expired segments as weak references,
allowing to spot "zombie" segments which should've been deleted by the time
we decide to trim them.
Since it's not a fatal error to have a dangling log segment given that it was
properly expired, the system will ignore zombies unless a dedicated new
debug flag is enabled:

- name: mds_debug_zombie_log_segments
  type: bool
  level: advanced
  default: false
  services:
  - mds
  fmt_desc: MDS will abort if an expired log segment is still
            strongly referenced (for developers only).

To achieve this it was necessary to change how segments are kept in the log.
Specifically, it was necessary to break up the `segments` map into two:
unexpired_segments and expired_segments, where the latter holds structs of
ExpiredSegmentInfo with a weak ref to the original log segment,
as well as some importany by-value copies for accounting and syncronization.

Since the change already required significant refactoring, it was used to deliver
a number of additional improvements:

* expiring segments don't need a separate container, they are now kept in `unexpired_segments`
  until they complete expiration. The expiration boundary is kept in a separate
  variable `last_expiring_segment_seq` which is sufficient.
* submit_mutex is now used exclusively to protect the `pending_events` when sending them
  to the submit thread. All other internal state is already protected by the mds_lock
* expiration waiters are now added as a single context associated with a segment sequence,
  which will be finished when no more unexpired segments exists with the sequence or below.
* expiration and trimming logic were optimized to make as much use of the O(log N) lookup
  in the std containers as possible.
* methods and fields were renamed to reduce ambiguity

Signed-off-by: Leonid Usov <leonid.usov@ibm.com>