cephadm: create ceph-exporter sock dir if it's not present
Since this is usually /var/run/ceph/ which ends up getting
created by other daemons as well, it was common to see
ceph-exporter fail to deploy and then deploy fine after
once other daemons were down on the host. I don't see any
reason we can't just try to make the directory here instead
of bailing out.

This patch had to be rewritten for reef, as it depended on
changes in cephadm that will not be backported to reef.

Fixes: https://tracker.ceph.com/issues/64491

Signed-off-by: Adam King <adking@redhat.com>