test/ceph_test_rados: fix seeking over virtual, in-memory objects
`ceph_test_rados` maintains a database of in-memory pseudo-objects
that are used to dynamically 1) generate content for writing into
RADOS objects and 2) regenerate it later when comparing buffers of
completed RADOS reads.

Content of these objects is ephemeral; to preserve space we store
only starting parameters for a pseudo-random number generator.

However, this memory optimization requires strict management of
calls to the PRGN. Unfortunately, adding partial reads support has
unveiled a bug in this aspect: `ObjectDesc::iterator::seek()` was
missing a call to underlying `RandGenerator::iterator_impl::seek()`,
resulting in mismatches on read buffer validation.

Signed-off-by: Radoslaw Zarzynski <rzarzyns@redhat.com>